Enable -Wsign-compare

+++ runtime/src/main/cpp/Porting.cpp
@@ -154,7 +154,7 @@ void consolePrintf(const char* format, ...) {
   va_start(args, format);
   int rv = vsnprintf_impl(buffer, sizeof(buffer), format, args);
   if (rv < 0) return; // TODO: this may be too much exotic, but should i try to print itoa(error) and terminate?
-  if (rv >= sizeof(buffer)) rv = sizeof(buffer) - 1;  // TODO: Consider realloc or report truncating.
+  if ((size_t)rv >= sizeof(buffer)) rv = sizeof(buffer) - 1;  // TODO: Consider realloc or report truncating.
   va_end(args);
   consoleWriteUtf8(buffer, rv);
 }
@@ -166,7 +166,7 @@ void consoleErrorf(const char* format, ...) {
   va_start(args, format);
   int rv = vsnprintf_impl(buffer, sizeof(buffer), format, args);
   if (rv < 0) return; // TODO: this may be too much exotic, but should i try to print itoa(error) and terminate?
-  if (rv >= sizeof(buffer)) rv = sizeof(buffer) - 1;  // TODO: Consider realloc or report truncating.
+  if ((size_t)rv >= sizeof(buffer)) rv = sizeof(buffer) - 1;  // TODO: Consider realloc or report truncating.
   va_end(args);
   consoleErrorUtf8(buffer, rv);
 }
